Renamed package and module name to remove rx
